Unify interface of structure comparison methods and `calculate_distance`
The new interface makes the second input of keys optional. If no key is provided, a similarity matrix will be constructed, to compare all pairs that can be constructed based on the first input. If the second input is provided, the lenghts need to match, since `keys1` and `keys2` are used for explicit pairwise comparison.